+++ app/src/main/java/com/scrappers/vitalwatch/data/cache/CacheQuickSetup.java
@@ -1,0 +1,4 @@
+package com.scrappers.vitalwatch.data.cache;
+
+public class CacheQuickSetup {
+}
