+++ app/src/main/java/com/scrappers/vitalwatch/data/cache/DataListener.java
@@ -1,0 +1,7 @@
+package com.scrappers.vitalwatch.data.cache;
+
+import com.scrappers.vitalwatch.data.SensorDataModel;
+
+public interface DataListener {
+    void onReadCompleted(final SensorDataModel cacheModel);
+}
